Extract Bedrock patch preflight helpers
